ajout infra webconfig
